fix: updated logging and cleanup of update methods

+++ internal/controller/nbnetworkresource/nbnetworkresource.go
@@ -18,7 +18,6 @@ package nbnetworkresource
 
 import (
 	"context"
-	"fmt"
 
 	xpv1 "github.com/crossplane/crossplane-runtime/apis/common/v1"
 	"github.com/crossplane/crossplane-runtime/pkg/connection"
@@ -33,6 +32,7 @@ import (
 	"github.com/crossplane/netbird-crossplane-provider/apis/vpn/v1alpha1"
 	nbcontrol "github.com/crossplane/netbird-crossplane-provider/internal/controller/nb"
 	"github.com/crossplane/netbird-crossplane-provider/internal/features"
+	"github.com/go-logr/logr"
 	netbird "github.com/netbirdio/netbird/management/client/rest"
 	nbapi "github.com/netbirdio/netbird/management/server/http/api"
 	"github.com/pkg/errors"
@@ -160,6 +160,7 @@ type external struct {
 	// A 'client' used to connect to the external resource API. In practice this
 	// would be something like an AWS SDK client.
 	service *NbService
+	log     logr.Logger
 }
 
 func (c *external) Observe(ctx context.Context, mg resource.Managed) (managed.ExternalObservation, error) {
@@ -167,13 +168,14 @@ func (c *external) Observe(ctx context.Context, mg resource.Managed) (managed.Ex
 	if !ok {
 		return managed.ExternalObservation{}, errors.New(errNotNbNetworkResource)
 	}
+	c.log.Info("observing", "cr", cr)
 	externalName := meta.GetExternalName(cr)
 	if externalName == "" {
 		return managed.ExternalObservation{ResourceExists: false}, nil
 	}
 	networks, err := c.service.nbCli.Networks.List(ctx)
 	if err != nil {
-		fmt.Printf("received error on call to nb: %+v", err)
+		c.log.Error(err, "received error on call to nb listing networks")
 		return managed.ExternalObservation{
 			ResourceExists: false,
 		}, nil //return nil so that observe can return without error so that it passes to create.
@@ -232,6 +234,7 @@ func (c *external) Create(ctx context.Context, mg resource.Managed) (managed.Ext
 	if !ok {
 		return managed.ExternalCreation{}, errors.New(errNotNbNetworkResource)
 	}
+	c.log.Info("creating", "cr", cr)
 	networks, err := c.service.nbCli.Networks.List(ctx)
 	if err != nil {
 		return managed.ExternalCreation{}, err
@@ -267,10 +270,8 @@ func (c *external) Create(ctx context.Context, mg resource.Managed) (managed.Ext
 	})
 
 	if err != nil {
-		fmt.Printf("err creating networkresource: %+v", err)
 		return managed.ExternalCreation{}, err
 	}
-	fmt.Printf("networkresource created: %+v", networkresource)
 	meta.SetExternalName(cr, networkresource.Id)
 	return managed.ExternalCreation{}, nil
 }
@@ -289,7 +290,7 @@ func (c *external) Update(ctx context.Context, mg resource.Managed) (managed.Ext
 		return managed.ExternalUpdate{}, errors.New(errNotNbNetworkResource)
 	}
 	networkResourceId := meta.GetExternalName(cr)
-	fmt.Printf("Updating: %+v", cr)
+	c.log.Info("Updating", "cr", cr)
 	networks, err := c.service.nbCli.Networks.List(ctx)
 	if err != nil {
 		return managed.ExternalUpdate{}, err
@@ -341,7 +342,7 @@ func (c *external) Delete(ctx context.Context, mg resource.Managed) error {
 		return errors.New(errNotNbNetworkResource)
 	}
 
-	fmt.Printf("Deleting: %+v", cr)
+	c.log.Info("Deleting", "cr", cr)
 	networks, err := c.service.nbCli.Networks.List(ctx)
 	if err != nil {
 		return err
@@ -351,6 +352,9 @@ func (c *external) Delete(ctx context.Context, mg resource.Managed) error {
 		if network.Name == cr.Spec.ForProvider.NetworkName {
 			apinetwork = &network
 		}
+	}
+	if apinetwork == nil {
+		return errors.New("network not found")
 	}
 	networkresourceid := meta.GetExternalName(cr)
 	return c.service.nbCli.Networks.Resources(apinetwork.Id).Delete(ctx, networkresourceid)
